+++ src/syngen/infer.py
@@ -5,6 +5,7 @@ import click
 from loguru import logger
 
 from syngen.ml.worker import Worker
+from syngen.ml.processors import PostprocessHandler
 from syngen.ml.utils import (
     setup_logger,
     create_log_file
@@ -122,6 +123,8 @@ def launch_infer(
     )
 
     worker.launch_infer()
+
+    PostprocessHandler(metadata_path, table_name, settings).run()
 
 
 if __name__ == "__main__":
