--- src/syngen/ml/preprocess/__init__.py
@@ -1,1 +1,0 @@
-from syngen.ml.preprocess.preprocess import PreprocessHandler  # noqa: F401
